[refactor] 레포지토리 인터페이스 변경
