+++ src/main/java/kr/hhplus/be/server/domain/payment/PaymentRepository.java
@@ -1,5 +1,8 @@
-package kr.hhplus.be.server.domain.order;
+package kr.hhplus.be.server.domain.payment;
+
+import java.util.Optional;
 
 public interface PaymentRepository {
     Payment save(Payment payment);
+    Optional<Payment> findById(Long id);
 }
