+++ src/main/java/kr/hhplus/be/server/domain/product/ProductRepository.java
@@ -1,8 +1,13 @@
 package kr.hhplus.be.server.domain.product;
 
+import kr.hhplus.be.server.domain.product.enums.SellingStatus;
+
 import java.util.List;
 
 public interface ProductRepository {
-    List<Product> findAllWithStock();
-    Product findById(Long id);
+    Product save(Product product);
+
+    Product findById(Long productId);
+
+    List<Product> findSellingStatusIn(List<SellingStatus> statuses);
 }
