+++ src/main/java/kr/hhplus/be/server/domain/rank/RankRepository.java
@@ -1,4 +1,10 @@
 package kr.hhplus.be.server.domain.rank;
 
-public class RankRepository {
+import java.util.List;
+
+public interface RankRepository {
+
+    Rank save(Rank rank);
+
+    List<RankInfo.PopularProduct> findPopularSellRanks(RankCommand.PopularSellRank command);
 }
