+++ src/main/java/kr/hhplus/be/server/domain/stock/StockRepository.java
@@ -1,4 +1,10 @@
 package kr.hhplus.be.server.domain.stock;
 
-public class StockRepository {
+public interface StockRepository {
+
+    Stock save(Stock stock);
+
+    Stock findByProductId(Long productId);
+
+    Stock findByProductIdWithLock(Long productId);
 }
